+++ API/src/main/java/fr/maxlego08/essentials/api/storage/IStorage.java
@@ -110,9 +110,8 @@ public interface IStorage {
      * Updates a user's money asynchronously.
      *
      * @param uniqueId the unique identifier of the user
-     * @param consumer the consumer to apply to the user
      */
-    void updateUserMoney(UUID uniqueId, Consumer<User> consumer);
+    User updateUserMoney(UUID uniqueId);
 
     /**
      * Retrieves the economy data of a user by their name.
@@ -599,4 +598,6 @@ public interface IStorage {
     StepDTO selectStep(UUID uniqueId, Step step);
 
     void registerStep(UUID uniqueId, Step step, String data);
+
+    List<String> getPlayerNames();
 }