+++ src/main/java/fr/maxlego08/essentials/commands/commands/utils/experience/CommandExperience.java
@@ -16,6 +16,7 @@ public class CommandExperience extends VCommand {
         this.addSubCommand(new CommandSetExperience(plugin));
         this.addSubCommand(new CommandQueryExperience(plugin));
         this.addSubCommand(new CommandGrantRandomExperience(plugin));
+        this.addSubCommand(new CommandTakeExperience(plugin));
     }
 
     @Override