+++ src/main/java/fr/maxlego08/essentials/migrations/CreateSanctionsTableMigration.java
@@ -1,6 +1,5 @@
 package fr.maxlego08.essentials.migrations;
 
-import fr.maxlego08.sarah.SchemaBuilder;
 import fr.maxlego08.sarah.database.Migration;
 
 public class CreateSanctionsTableMigration extends Migration {
@@ -13,7 +12,7 @@ public class CreateSanctionsTableMigration extends Migration {
             table.string("sanction_type", 255);
             table.longText("reason");
             table.bigInt("duration");
-            table.timestamp("expired_at").nullable();
+            table.timestamp("expired_at");
             table.timestamps();
         });
     }